requirements.txt
fastapi==0.112.0
uvicorn==0.30.3
python-dotenv==1.0.1
asyncpg==0.29.0
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
redis==5.0.7
rq==1.16.2
httpx==0.27.0
EOF
